refactor: move tooltip and sheet-specific trigger buttons to those directories
